move fioaction_object to its own file
